Make work_division._matmul_split_cost symbolic aware
